Checkout detectSystem utils
