+++ src/components/utils/index.js
@@ -2,31 +2,6 @@ import parseUnit from 'parse-unit'
 import { h } from 'preact'
 import enumerateDevices from 'enumerate-devices'
 
-const SYSTEMS_MAPPING = {
-  os: [
-    { name: 'Windows Phone', codeName: 'Windows Phone', versionPrefix: 'OS' },
-    { name: 'Windows', codeName: 'Win', versionPrefix: 'NT' },
-    { name: 'iPhone', codeName: 'iPhone', versionPrefix: 'OS' },
-    { name: 'iPad', codeName: 'iPad', versionPrefix: 'OS' },
-    { name: 'Kindle', codeName: 'Silk', versionPrefix: 'Silk' },
-    { name: 'Android', codeName: 'Android', versionPrefix: 'Android' },
-    { name: 'PlayBook', codeName: 'PlayBook', versionPrefix: 'OS' },
-    { name: 'BlackBerry', codeName: 'BlackBerry', versionPrefix: '/' },
-    { name: 'Macintosh', codeName: 'Mac', versionPrefix: 'OS X' },
-    { name: 'Linux', codeName: 'Linux', versionPrefix: 'rv' },
-    { name: 'Palm', codeName: 'Palm', versionPrefix: 'PalmOS' },
-  ],
-  browser: [
-    { name: 'Chrome', codeName: 'Chrome', versionPrefix: 'Chrome' },
-    { name: 'Firefox', codeName: 'Firefox', versionPrefix: 'Firefox' },
-    { name: 'Safari', codeName: 'Safari', versionPrefix: 'Version' },
-    { name: 'Internet Explorer', codeName: 'MSIE', versionPrefix: 'MSIE' },
-    { name: 'Opera', codeName: 'Opera', versionPrefix: 'Opera' },
-    { name: 'BlackBerry', codeName: 'CLDC', versionPrefix: 'CLDC' },
-    { name: 'Mozilla', codeName: 'Mozilla', versionPrefix: 'Mozilla' },
-  ],
-}
-
 export const functionalSwitch = (key, hash) => (hash[key] || (() => null))()
 
 export const getCSSValue = (expectedUnit, cssUnit) => {
@@ -166,40 +141,3 @@ export const addDeviceRelatedProperties = (sdkMetadata, isCrossDeviceFlow) => ({
   isCrossDeviceFlow,
   deviceType: isDesktop ? 'desktop' : 'mobile',
 })
-
-export const extractSystemInfo = (systemType) => {
-  const userAgent = [
-    window.navigator.platform,
-    window.navigator.userAgent,
-    window.navigator.appVersion,
-    window.navigator.vendor,
-    window.opera,
-  ].join(' ')
-
-  const matchedSystems = SYSTEMS_MAPPING[systemType].map(
-    ({ codeName, name, versionPrefix }) => {
-      const codeNameRegex = new RegExp(codeName, 'gi')
-
-      // No system codeNames found
-      if (!codeNameRegex.test(userAgent)) {
-        return null
-      }
-
-      const versionRegex = new RegExp(
-        `${versionPrefix}[-\\s\\/:;]([\\d\\._]+)`,
-        'i'
-      )
-      const matches = userAgent.match(versionRegex)
-      const version =
-        matches && matches[1] ? matches[1].split(/[._]+/).join('.') : '0'
-
-      return {
-        name,
-        version,
-      }
-    }
-  )
-
-  // Get the first non-nullish object
-  return matchedSystems.filter((el) => el)[0]
-}
